+++ dto/repository/inbound/extend_inbound.go
@@ -4,5 +4,5 @@ import "time"
 
 type ExtendInboundDto struct {
 	End             time.Time
-	TrafficExtended uint32
+	TrafficExtended uint64
 }
